% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.rrotsfspls.R
\name{predict.rrotsfspls}
\alias{predict.rrotsfspls}
\title{rotsf predict.rrotsfspls}
\usage{
\method{predict}{rrotsfspls}(object, testx, trlength = object$trlength, ...)
}
\arguments{
\item{object}{An object that inherits from class rrotsfspls.}

\item{testx}{A data frame in which to look for variables with which to predict.}

\item{trlength}{Number of based models used for prediction, shouble be less than and equal to the number for training.}

\item{...}{Additional arguments.}
}
\value{
produces a vector of predictions or a matrix of predictions
}
\description{
Prediction with rrotsfspls and return a list of with mean hazard function
}
\examples{
set.seed(123)
require(rotsf)
require(survival)
## Survival Forest with PLS  with default settings
#Lung DATA
data(lung)
lung=na.omit(lung)
lung[,3]=lung[,3]-1
n=dim(lung)[1]
L=sample(1:n,ceiling(n*0.5))
trset<-lung[L,]
teset<-lung[-L,]
rii=c(2,3)
plssurvmodel=rrotsfspls(x=trset[,-rii],y=Surv(trset[,rii[1]], trset[,rii[2]]))
mypre=predict(plssurvmodel, teset[,-rii])
}
\references{
\itemize{
  \item Zhou L, Xu Q, Wang H. (2015) Rotation survival forest for right censored data. PeerJ 3:e1009 https://doi.org/10.7717/peerj.1009.
  \item Zhou, L., Wang, H., & Xu, Q. (2016). Random rotation survival forest for high dimensional censored data. SpringerPlus, 5(1), 1425.
 }
}
\seealso{
\code{pec} R package
}
\author{
Hong Wang and Lifeng Zhou
}
