\name{rotsf-package}
\alias{rotsf-package}
\alias{rotsf}
\docType{package}
\title{
Rotation survival forest for right censored data
}
\description{
Rotation survival forest for right censored data
}
\details{
\tabular{ll}{
Package: \tab rotsf\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-03-23\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
Zhou L, Xu Q, Wang H. (2015) Rotation survival forest for right censored data. PeerJ 3:e1009 https://doi.org/10.7717/peerj.1009
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
