% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotsfpca_predict.R
\name{rotsfpca.predict}
\alias{rotsfpca.predict}
\title{Prediction with new data and return a saved forest with mean hazard function}
\usage{
rotsfpca.predict(rotsffit, newdata, trlength = 500)
}
\description{
Prediction with new data and return a saved forest with mean hazard function
}
