% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotsfpca_surv_predict.R
\name{rotsfpca.surv_predict}
\alias{rotsfpca.surv_predict}
\title{Prediction with new data and return a saved forest with survival proability at.
Each unique time points}
\usage{
rotsfpca.surv_predict(rotsfpcafit, newdata, uniquetimes, trlength = 500)
}
\description{
Prediction with new data and return a saved forest with survival proability at.
Each unique time points
}
