% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrotsfpls_surv.predict.R
\name{rrotsfpls_surv.predict}
\alias{rrotsfpls_surv.predict}
\title{Prediction with new data and return a saved forest with mean hazard function}
\usage{
rrotsfpls_surv.predict(object, newdata, uniquetimes,
  trlength = object$trlength)
}
\description{
Prediction with new data and return a saved forest with mean hazard function
}
