% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrotsfpls.R
\name{rrotsfspls}
\alias{rrotsfspls}
\title{rotsf rrotsfspls}
\usage{
rrotsfspls(x, y, trlength = 100, mtry = floor(sqrt(ncol(x))),
  impute = TRUE, nplscomp = floor(sqrt(ncol(x))), vari_status = FALSE,
  ...)
}
\arguments{
\item{x}{The covariates(predictor variables) of training data.}

\item{y}{Survival time and censored status of training data. Must be a Surv  \code{survival} object}

\item{trlength}{The ensemle size (the number of base  survival trees). Default is 100.}

\item{mtry}{The number of covariates(predictor variables) used in each base tree model. Default is the square root of the number of all avaibable covariates.}

\item{vari_status}{Whether or not calculate variables importance scores. Default is "FALSE".}

\item{...}{Additional arguments for the base decision tree, see the \code{rpart} package for details.}
}
\value{
Object of class \code{rrotsfspls} with elements
  \tabular{ll}{
      \code{pectrees}    \tab  A list of base models \code{pecRpart} in \code{pec} R package of size \code{trlength}. To retrieve a particular base model: use  pectrees[[i]], where i takes values between 1 and \code{trlength} \cr
      \code{colindexes} \tab A list of covaraite subspace index for each base tree. \cr
      \code{trlength} \tab Number of bases models trained. \cr
      \code{rotms} \tab A list of PLS weights  matrix.To retrieve a particular weight matrix: use  rotms[[i]], where i takes values between 1 and \code{trlength} \cr
      \code{varimp} \tab If  \code{vari_status=FALSE}, return a Matrix of variable importance scores. \cr
  }
}
\description{
Grow a survival forest with partial least squares(PLS)
}
\examples{
set.seed(123)
require(rotsf)
require(survival)
## Survival Forest with PLS  with default settings
#Lung DATA
data(lung)
lung=na.omit(lung)
lung[,3]=lung[,3]-1
n=dim(lung)[1]
L=sample(1:n,ceiling(n*0.5))
trset<-lung[L,]
teset<-lung[-L,]
rii=c(2,3)
plssurvmodel=rrotsfspls(x=trset[,-rii],y=Surv(trset[,rii[1]], trset[,rii[2]]))
# Get the 1th base model
firstbasemodel=plssurvmodel$pectrees[[1]]
#second PLS weight matrix
secondweigmatrix=plssurvmodel$rotms[[2]]
plssurvmodel2=rrotsfspls(x=trset[,-rii],y=Surv(trset[,rii[1]], trset[,rii[2]]),vari_status=TRUE)
#variable importance
varimpscores=plssurvmodel2$varimp
}
\references{
\itemize{
  \item Zhou L, Xu Q, Wang H. (2015) Rotation survival forest for right censored data. PeerJ 3:e1009 https://doi.org/10.7717/peerj.1009.
  \item Zhou, L., Wang, H., & Xu, Q. (2016). Random rotation survival forest for high dimensional censored data. SpringerPlus, 5(1), 1425.
 }
}
\seealso{
\code{pec} R package
}
\author{
Hong Wang and Lifeng Zhou
}
